liquify: check for handles using min distance
The previous algorithm was using the cairo routine to check to intersection.
This was not good when control points where close to each others. The
new implementation separate the drawing of the forms and the check
for closest handle near the mouse.

This new implementation makes the algorithm clearer and simplify the
code and at the same time gives easier management of the control
points.