histogram: handle CLUT display profiles for vectorscope
Remove optimization. Put back in a profile transform hop and temporary
buffer to hold colorspace transform result in the case of vectorscope.
The optimization code could only handle matrix profiles, and as the
input data is in display profile, which may be CLUT, this code broke
in that case.

Essentially reverts 955ee4b3e7136454ffc555539aa8cbfd134f1a20.